requirements.txt
pillow==11.3.0
tkfontawesome==0.3.2
colorlog==6.9.0
pytest==8.4.2
pytest-mock==3.15.1